mqtt_reader/requirements.txt
async-timeout==4.0.2
Deprecated==1.2.13
packaging==21.3
paho-mqtt==1.6.1
pyparsing==3.0.9
python-dotenv==0.21.0
redis==4.3.4
wrapt==1.14.1
